Fixes #4106 

From my testing (printing out elapsed `System.getCurrentTime()` before and after), it has reduced the average method call time from 6000 ms to 2200 ms.

To test:

Start up the app cleanly (removed from the back stack) both when signed in and signed out. Notice how much faster it is when you are already signed in.

Needs review: @maxme