@maxme 2016-05-20
So I've run 20 tests with the 2 configurations (login screen vs. logged in user):

* login screen: from 281ms to 2757ms startup time, median: 1399ms
* logged in: from 351ms to 839ms startup time: median 570ms

So definitely better when the user is logged in.

For sure, there is a blocking network call when initializing with `startOptimizelyWithAPIToken` with a quick timeout (I tried to suspend the call with a proxy).

We must kill the blocking init when the login screen (magic link) AB test is finished.