@maxme 2016-05-20
We should contact Optimizely, or open a bug report, I see some OkHttp warnings:
```
05-20 11:31:33.302 14302-14434/org.wordpress.android.beta W/okhttp3.OkHttpClient: A connection to http://5674220492.log.optimizely.com/ was leaked. Did you forget to close a response body?
05-20 11:31:33.302 14302-14434/org.wordpress.android.beta W/okhttp3.OkHttpClient: A connection to http://5674220492.log.optimizely.com/ was leaked. Did you forget to close a response body?
05-20 11:32:50.157 18146-18289/org.wordpress.android.beta W/okhttp3.OkHttpClient: A connection to http://5674220492.log.optimizely.com/ was leaked. Did you forget to close a response body?
05-20 11:34:17.863 20982-21103/org.wordpress.android.beta W/okhttp3.OkHttpClient: A connection to http://5674220492.log.optimizely.com/ was leaked. Did you forget to close a response body?
```